Added scheduled function to check admin lists
- FB scheduled function will compare author doc admin list against newly-fetched list from Github API, and update if necessary; runs every 10 minutes
- reorganized/cleaned up GH utility functions in cloud function folders
- minor updates to types/interfaces
