+++ functions/src/fb-event-triggers/auth.ts
@@ -1,8 +1,8 @@
 import * as functions from "firebase-functions";
 import admin, { db } from "../config/fbConfig";
 
-import { getGithubUsername } from "../utils";
-import { User } from "../types";
+import { getGithubUsername } from "../utils/github";
+import { User } from "../../../types/auth";
 
 // cloud function to create user profile upon signin for the first time
 export const createUser = functions.auth.user().onCreate(async (user: admin.auth.UserRecord) => {
@@ -10,7 +10,8 @@ export const createUser = functions.auth.user().onCreate(async (user: admin.auth
     name: user.displayName || null,
     email: user.email || null,
     roles: ["user"],
-    uid: user.uid
+    uid: user.uid,
+    config: []
   };
   // If provider is GitHub, and user has no public name, use username for name
   if (user.providerData[0].providerId === "github.com" && !user.displayName) {
