+++ types/author.d.ts
@@ -15,6 +15,7 @@ interface AuthorSourceConfig {
 interface GithubOrgConfig {
   id: number;
   name: string;
+  type: "User" | "Organization";
   installed: boolean;
   installationId: number;
   admins: number[];
